fix(codex): harden tool argument normalization
